refactor: state값이 변경되면 useEffect로 로컬스토리지 값 저장

+++ src/pages/Detail/DetailForm.jsx
@@ -102,6 +102,8 @@ const DetailForm = ({ post = {}, onUpdate, onDelete }) => {
         <Button
           variant="destructive"
           onClick={() => {
+            if (!confirm("삭제하실건가요? 🥲")) return;
+
             navigate(-1);
             onDelete(post.id);
           }}
